Added Dockerfile
Added the Dockerfile to the maven-web-project. (docker run -i maven-web-project) to run the image.

However it does not work as it says it is missing a jar file. IDK how to fix that so im just gonna leave it as it is right now. Im good with what it is currently.